swap unittest with pytest for better discovery of test and refactoring of webapp element
