+++ ai4good/webapp/input_page_3.py
@@ -8,6 +8,7 @@ from dash.dependencies import Input, Output, State, MATCH, ALL
 from dash.exceptions import PreventUpdate
 from ai4good.webapp.apps import dash_app, facade, model_runner, _redis
 import ai4good.webapp.run_model_page as run_model_page
+import ai4good.webapp.common_elements as common_elements
 from ai4good.webapp.model_runner import InputParameterCache
 import ai4good.utils.path_utils as pu
 
@@ -17,7 +18,7 @@ radio_option_3 = [{'label':'Yes', 'value':1}, {'label':'No', 'value':0}, {'label
 
 layout = html.Div(
     [
-        run_model_page.nav_bar(),
+        common_elements.nav_bar(),
  
         html.Div([
             dbc.Container([
