+++ ai4good/webapp/landing_page.py
@@ -1,20 +1,13 @@
-import numpy as np
-import pandas as pd
-import dash
-import dash_core_components as dcc
 import dash_bootstrap_components as dbc
 import dash_html_components as html
-from ai4good.webapp.apps import dash_app, facade, model_runner
-import ai4good.webapp.run_model_page as run_model_page
-from ai4good.webapp.model_runner import ModelScheduleRunResult
-from dash.dependencies import Input, Output, State
-import dash_table
+import ai4good.webapp.common_elements as common_elements
+
 
 text_introduction = 'The Simulator is a web tool for NGOs and local authorities to model COVID-19 outbreak inside refugee camps and prepare timely and proportionate response measures needed to flatten the curve and reduce the number of fatalities. This tool helps to predict the possble outbreak scenarios and their potential outcomes and help NGOs design an optimal intervention strategy. '
 
 layout = html.Div(
     [
-        run_model_page.nav_bar(),
+        common_elements.nav_bar(),
         
         html.Div([
             dbc.Container([
