+++ fs/params/compartmental-model_profile_params.csv
@@ -71,3 +71,12 @@ combined_hygiene_symptomatic_high_risk,remove_high_risk_categories,<no_edit>,<no
 combined_hygiene_symptomatic_high_risk,t_sim,<no_edit>,<no_edit>,200
 combined_hygiene_symptomatic_high_risk,numberOfIterations,<no_edit>,<no_edit>,1000
 combined_hygiene_symptomatic_high_risk,nProcesses,<no_edit>,<no_edit>,4
+custom,better_hygiene,0,200,0.7
+custom,ICU_capacity,<no_edit>,<no_edit>,6
+custom,remove_symptomatic,30,90,50
+custom,shielding,<no_edit>,<no_edit>,False
+custom,remove_high_risk,0,12,50
+custom,remove_high_risk_categories,<no_edit>,<no_edit>,2
+custom,t_sim,<no_edit>,<no_edit>,200
+custom,numberOfIterations,<no_edit>,<no_edit>,1000
+custom,nProcesses,<no_edit>,<no_edit>,4
