+++ run_tests.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-python -m unittest discover -s ai4good/ -p "test_*.py"
+pytest ./ai4good --cov=./ --cov-report=xml
